requirements.txt
beautifulsoup4
selenium
discord
Flask
flask-restful
python-dotenv
psycopg2-binary
pandas
SQLAlchemy